requirements.txt
PyQt6==6.6.1
pymysql==1.1.0
bcrypt==4.1.2
python-dotenv==1.0.0
reportlab==4.0.9
openpyxl==3.1.2
matplotlib==3.8.2
pyinstaller==6.3.0